Alternative to ModioFilterCreator: filter strings
Bypass ModioFilterCreator to ease filter setup.
Takes the actual query string for the API call underneath.

Adds *FilterString() variants for every function originally
requiring a ModioFilterCreator struct.

e.g. build a query string in script, push it to C and call modioGetAllModsFilterString(), done.
Without this bypass the script generates a query string (or something else) to push the information to C, in C the data is rippped apart to construct ModioFilterCreator, which later concatenates all this data back into a query string. Quite a round-trip..